pyproject.toml
[tool.poetry]
name = "white-square-maker"
version = "0.1.0"
description = ""
authors = ["Patrick Rice <patrickrice20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
